move usePlaces to it's own file
