+++ src/components/Place/PlaceTable.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { usePlaces } from "../Places/Places";
+import { usePlaces } from "../../data/model/PlaceModel";
 import { useFirebase } from '../../util/Firebase';
 import { useAuth } from "../../util/UserAuth";
