+++ src/components/Places/Places.js
@@ -7,6 +7,7 @@ import ModalWrapper from "../Modal/Modal";
 import { useAuth } from "../../util/UserAuth";
 import MPlaceTable from "../Place/MPlaceTable";
 import { ProfileBtn } from "../Profile/Profile";
+import { usePlaces } from "../../data/model/PlaceModel";
 
 import Fab from '@material-ui/core/Fab';
 import Icon from '@material-ui/core/Icon';
@@ -121,49 +122,4 @@ const MapButtons = ({onClickAdd, onClickList}) => {
   );
 }
 
-/**
- * Place hook, gives the users places in a format that is
- * expected by the rest of the app.
- */
-const usePlaces = () => {
-  // TODO: Init as null or empty list?
-  const [places, setPlaces] = useState(null);
-
-  const auth = useAuth();
-  const firebase = useFirebase();
-
-  /**
-   * Given a Firebase snapshot of place it returns an array of 
-   * place objects that the webapp can work with.
-   * @param {Object} snapshot snapshot from firebase
-   * @return {Array} List of place objects 
-   */
-  const convert = (snapshot) => {
-    let keys = Object.keys(snapshot.val());
-    let firebasePlaces = Object.values(snapshot.val());
-    firebasePlaces.forEach((place, idx) => {
-      place.id = keys[idx];             // Add key as object id
-      place.date = new Date(place.date);    // Convert date to object
-    });
-    return firebasePlaces;
-  }
-
-  // Get user place from fierbase
-  useEffect(() => {
-    if (auth.user) {
-      firebase.place(auth.user.uid).on('value', (snapshot) => {
-        if (snapshot.val()) {
-          setPlaces(convert(snapshot));
-        }
-      });
-    } else {
-      setPlaces([]);
-    }
-  }, [auth, firebase]);
-
-  return places;
-}
-
 export default Places;
-
-export { usePlaces }
